+++ inkcpp_compiler/binary_emitter.cpp
@@ -12,7 +12,6 @@
 
 #include <vector>
 #include <map>
-#include <fstream>
 
 #ifndef _MSC_VER
 #	include <cstring>
@@ -59,7 +58,7 @@ struct container_data {
 	uint32_t end_offset = 0;
 
 	// Index used in CNT? operations
-	container_t counter_index = ~0;
+	container_t counter_index = ~0U;
 
 	~container_data()
 	{
@@ -240,7 +239,7 @@ void binary_emitter::output(std::ostream& out)
 	write_container_map(out, _container_map, _max_container_index);
 
 	// Write a separator
-	uint32_t END_MARKER = ~0;
+	uint32_t END_MARKER = ~0U;
 	out.write(( const char* ) &END_MARKER, sizeof(uint32_t));
 
 	// Write container hash list
@@ -320,7 +319,7 @@ void binary_emitter::process_paths()
 		bool firstParent = true;
 
 		// We need to parse the path
-		offset_t    noop_offset = ~0;
+		offset_t    noop_offset = ~0U;
 		char*       _context    = nullptr;
 		const char* token
 		    = ink::compiler::internal::strtok_s(const_cast<char*>(path_cstr), ".", &_context);
@@ -367,7 +366,7 @@ void binary_emitter::process_paths()
 		} else {
 			// If we want the count index, write that out
 			if (useCountIndex) {
-				inkAssert(container->counter_index != ~0, "No count index available for this container!");
+				inkAssert(container->counter_index != ~0U, "No count index available for this container!");
 				_containers.set(position, container->counter_index);
 			} else {
 				// Otherwise, write container address
@@ -435,11 +434,17 @@ void binary_emitter::set_list_meta(const list_data& list_defs)
 		_lists.write(flag.flag);
 		if (flag.flag.list_id != list_id) {
 			list_id = flag.flag.list_id;
-			_lists.write(reinterpret_cast<const byte_t*>(list_names->data()), list_names->size());
+			_lists.write(
+			    reinterpret_cast<const byte_t*>(list_names->data()),
+			    static_cast<size_t>(list_names->size())
+			);
 			++list_names;
 			_lists.write('\0');
 		}
-		_lists.write(reinterpret_cast<const byte_t*>(flag.name->c_str()), flag.name->size() + 1);
+		_lists.write(
+		    reinterpret_cast<const byte_t*>(flag.name->c_str()),
+		    static_cast<size_t>(flag.name->size()) + 1
+		);
 	}
 	_lists.write(null_flag);
 }
